requirements.txt
google-genai
google.genai